requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6

ddgs==9.7.1
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
hyperframe==6.1.0

# Web scraping enhancements
aiohttp==3.9.5
fake-useragent==2.0.0
tqdm==4.67.0
lxml==5.2.2
python-dotenv==1.0.1
urllib3==2.2.2
soupsieve==2.5

# Async support
asyncio==3.4.3
nest-asyncio==1.6.0

# Data processing
pandas==2.2.1
numpy==1.26.4

# Caching and rate limiting
cachetools==5.3.3
ratelimit==2.2.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==2.7.1
packaging==25.0
pillow==12.0.0
primp==0.15.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
six==1.17.0
smolagents==1.22.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
tqdm==4.67.1
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
